@@ -7,14 +7,18 @@ package com.draabek.androsigner;
 
 public class Constants {
     public static final String INTENT_RESULT_DESCRIPTION = "com.draabek.androsigner.RESULT_ACTION";
+    public static final String CONFIRM_REQUEST_ACTION = "com.draabek.androsigner.CONFIRM_REQUEST_ACTION";
     public static final String COMMAND_GENERATE_ADDRESS = "COMMAND_GENERATE_ADDRESS";
     public static final String COMMAND_CONFIRM_TRANSACTION = "COMMAND_CONFIRM_TRANSACTION";
     public static final String COMMAND_SIGN_MESSAGE = "COMMAND_SIGN_MESSAGE";
     public static final String COMMAND_LIST_ADDRESSES = "COMMAND_LIST_ADDRESSES";
     public static final String COMMAND_GET_BALANCE = "COMMAND_GET_BALANCE";
+    public static final String COMMAND_GET_PUBKEY = "COMMAND_GET_PUBKEY";
+    public static final String COMMAND_CONFIG = "COMMAND_CONFIG";
     public static final String RETURN_GENERATED_ADDRESS = "RETURN_GENERATED_ADDRESS";
+    public static final String RETURN_ADDRESS_MNEMONIC = "RETURN_ADDRESS_MNEMONIC";
     public static final String RETURN_TRANSACTION_HASH = "RETURN_TRANSACTION_HASH";
-    public static final String INTENT_KEY_COMMAND = "INTENT_KEY_COMMAND";
+    public static final String INTENT_COMMAND = "INTENT_COMMAND";
     public static final String INTENT_PASSWORD = "INTENT_PASSWORD";
     public static final String INTENT_FROM = "INTENT_FROM";
     public static final String INTENT_TO = "INTENT_TO";
@@ -23,6 +27,15 @@ public class Constants {
     public static final String INTENT_GAS_PRICE = "INTENT_GAS_PRICE";
     public static final String INTENT_GAS_LIMIT = "INTENT_GAS_LIMIT";
     public static final String INTENT_FAILURE_REASON = "INTENT_FAILURE_REASON";
-    public static final String INTENT_FAILURE_REASON_CANCELLED = "INTENT_FAILURE_REASON_CANCELLED";
+    public static final String INTENT_FAILURE_REASON_CANCELLED = "Request cancelled by user";
+    public static final String INTENT_FAILURE_REASON_BAD_INPUT = "Bad input received with request";
+    public static final String INTENT_FAILURE_INSUFFICIENT_GAS = "Insufficient gas";
+    public static final String INTENT_FAILURE_REASON_UNKNOWN_ERROR = "Unknown error";
+    public static final String COMMAND_CONFIRM_TRANSFER = "COMMAND_CONFIRM_TRANSFER";
+    public static final String INTENT_ACCOUNT = "INTENT_ACCOUNT";
+    public static final String RETURN_ACCOUNT_LIST = "RETURN_ACCOUNT_LIST";
+    public static final String INTENT_MESSAGE = "INTENT_MESSAGE";
+    public static final String INTENT_MESSAGE_SIGNATURE = "INTENT_MESSAGE_SIGNATURE";
+    public static final String INTENT_ACCOUNT_BALANCE = "INTENT_ACCOUNT_BALANCE";
 }
 
